pyproject.toml
[project]
name = "hw-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "envyaml>=1.10.211231",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langfuse>=3.2.0",
    "langgraph>=0.5.1",
    "langgraph-supervisor>=0.0.27",
    "langsmith>=0.4.4",
    "oci>=2.155.1",
    "python-box>=7.3.2",
    "python-dotenv>=1.1.1",
    "transformers>=4.53.1",
]
